modify process of bot. this is managed by supervisor
