+++ mix.exs
@@ -3,7 +3,7 @@ defmodule MagicBot.Mixfile do
 
   def project do
     [app: :magic_bot,
-     version: "0.0.1",
+     version: "0.0.2",
      elixir: "~> 1.0",
      build_embedded: Mix.env == :prod,
      start_permanent: Mix.env == :prod,
@@ -14,7 +14,8 @@ defmodule MagicBot.Mixfile do
   #
   # Type `mix help compile.app` for more information
   def application do
-    [applications: [:logger, :slack]]
+    [applications: [:logger, :slack],
+    mod: {MagicBot, []}]
   end
 
   # Dependencies can be Hex packages:
